fraud_detection_project/backend/requirement.txt
flask
flask-cors
pandas
joblib
scikit-learn
mysql-connector-python
numpy
